:truck: renaming: move `smp` to `smp_call`
