+++ components/drivers/smp_call/SConscript
