+++ examples/utest/testcases/Kconfig
@@ -15,7 +15,7 @@ rsource "drivers/ipc/Kconfig"
 rsource "posix/Kconfig"
 rsource "mm/Kconfig"
 rsource "tmpfs/Kconfig"
-rsource "smp_ipi/Kconfig"
+rsource "smp_call/Kconfig"
 endif
 
 endmenu
