+++ examples/utest/testcases/smp_call/Kconfig
