+++ examples/utest/testcases/smp_call/SConscript
