+++ libcpu/aarch64/common/setup.c
@@ -13,7 +13,7 @@
 #define DBG_TAG "cpu.aa64"
 #define DBG_LVL DBG_INFO
 #include <rtdbg.h>
-#include <smp.h>
+#include <smp_call.h>
 #include <cpu.h>
 #include <mmu.h>
 #include <cpuport.h>
